CERT rrToWire wrote keyTag incorrectly
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@197 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ DNS/CERTRecord.java
@@ -95,7 +95,7 @@ rrToWire(DataByteOutputStream dbs, Compression c) throws IOException {
 		return;
 
 	dbs.writeShort(certType);
-	dbs.writeByte(keyTag);
+	dbs.writeShort(keyTag);
 	dbs.writeByte(alg);
 	dbs.write(cert);
 }
